+++ src/common/io.h
@@ -9,9 +9,9 @@
 #define XGBOOST_COMMON_IO_H_
 
 #include <dmlc/io.h>
+#include <rabit/rabit.h>
 #include <string>
 #include <cstring>
-#include "./sync.h"
 
 namespace xgboost {
 namespace common {
