+++ src/tree/updater_basemaker-inl.h
@@ -7,6 +7,8 @@
 #ifndef XGBOOST_TREE_UPDATER_BASEMAKER_INL_H_
 #define XGBOOST_TREE_UPDATER_BASEMAKER_INL_H_
 
+#include <rabit/rabit.h>
+
 #include <xgboost/base.h>
 #include <xgboost/tree_updater.h>
 #include <vector>
@@ -14,8 +16,8 @@
 #include <string>
 #include <limits>
 #include <utility>
+
 #include "./param.h"
-#include "../common/sync.h"
 #include "../common/io.h"
 #include "../common/random.h"
 #include "../common/quantile.h"
